--- Project1/LinearSearch.h
@@ -1,1 +1,0 @@
-#pragma once
