--- Project1/QuickSort.h
@@ -1,25 +1,0 @@
-#pragma once
-#if 0
-#include <iostream>
-#include "LinkedList.h"
-
-using namespace std;
-
-struct Node {
-	string words;
-	Node* nextAddress;
-	Node(string word_list) : words(word_list), nextAddress(nullptr) {}
-};
-
-class QuickSort
-{
-private:
-	LinkedList linkedList;
-	Node* partition(Node* start, Node* end, Node** newHead, Node** newEnd);
-
-public:
-	QuickSort();
-	void sort(Node* start, Node* end);
-};
-
-#endif
