+++ Project1/x64/Debug/Project1.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\x64\Debug\Project1.exe</FullPath>
+      <FullPath>C:\Users\yizun\OneDrive\Documents\GitHub\Data_Structure_GA_2402\x64\Debug\Project1.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
