+++ Project1/x64/Debug/Project1.tlog/Cl.items.tlog
@@ -1,9 +1,2 @@
-C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\Array.cpp;C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\x64\Debug\Array.obj
-C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\FileHandler.cpp;C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\x64\Debug\FileHandler.obj
-C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\FileHandler.h;C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\x64\Debug\FileHandler.obj
-C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\LinearSearch.cpp;C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\x64\Debug\LinearSearch.obj
-C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\LinkedList.cpp;C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\x64\Debug\LinkedList.obj
-C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\main.cpp;C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\x64\Debug\main.obj
-C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\QuickSort.cpp;C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\x64\Debug\QuickSort.obj
-C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\WordLinkedList.cpp;C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\x64\Debug\WordLinkedList.obj
-C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\WordLinkedList.h;C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\Project1\x64\Debug\WordLinkedList.obj
+C:\Users\yizun\OneDrive\Documents\GitHub\Data_Structure_GA_2402\Project1\FileIO.h;C:\Users\yizun\OneDrive\Documents\GitHub\Data_Structure_GA_2402\Project1\x64\Debug\FileIO.obj
+C:\Users\yizun\OneDrive\Documents\GitHub\Data_Structure_GA_2402\Project1\main.cpp;C:\Users\yizun\OneDrive\Documents\GitHub\Data_Structure_GA_2402\Project1\x64\Debug\main.obj
