+++ Project1/x64/Debug/Project1.tlog/Project1.lastbuildstate
@@ -1,2 +1,2 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.40.33807:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|C:\Users\Predator\OneDrive - Asia Pacific University\Current Assignments\DSTR\DSTR Part 1 Code\Data_Structure_GA_2402\|
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.41.34120:TargetPlatformVersion=10.0.22621.0:
+Debug|x64|C:\Users\yizun\OneDrive\Documents\GitHub\Data_Structure_GA_2402\|
